# Optimize label values with matchers by taking shortcuts