We shouldn't calculate postings beforehand: we may not need them.
If all matchers are for the requested label, we can just filter its values.

This is the common case of someone asking for metrics autocomplete by calling `/label/__name__/values?match={__name__="foo.*"}`.

And also, if there are no values at all, there's no need to run any kind of logic.

I also added some extra benchmarks. Benchmarks take a while, so I only ran them for `Head`. They should look the same for `Block`:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
                                                                        │       main       │                  new                  │
                                                                        │      sec/op      │    sec/op     vs base                 │
Querier/Head/labelValuesWithMatchers/i_with_i="1"-12                          1.176m ±  2%   1.162m ±  6%         ~ (p=0.280 n=10)
Querier/Head/labelValuesWithMatchers/i_with_n="1"-12                          93.09m ± 13%   85.45m ±  4%    -8.21% (p=0.043 n=10)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"-12                      148.97m ±  1%   42.59m ±  3%   -71.41% (p=0.000 n=10)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"-12                 101.2m ± 11%   101.8m ±  4%         ~ (p=0.853 n=10)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j=~"X.+"-12                 14.04m ±  3%   13.77m ±  3%         ~ (p=0.123 n=10)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j=~"XXX|YYY"-12             14.08m ±  2%   13.84m ±  2%    -1.67% (p=0.043 n=10)
Querier/Head/labelValuesWithMatchers/i_with_n="X",j!="foo"-12                 75.91m ±  6%   74.10m ± 17%         ~ (p=0.143 n=10)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"-12       185.5m ±  5%   181.1m ±  6%    -2.33% (p=0.035 n=10)
Querier/Head/labelValuesWithMatchers/i_with_i="1aaa...ddd"-12                 1.205m ±  6%   1.186m ± 13%         ~ (p=0.739 n=10)
Querier/Head/labelValuesWithMatchers/i_with_i=~"1.+"-12                      18.500m ± 33%   6.132m ± 19%   -66.85% (p=0.000 n=10)
Querier/Head/labelValuesWithMatchers/i_with_i=~"1.+",i=~".+ddd"-12            75.71m ± 10%   10.12m ±  1%   -86.63% (p=0.000 n=10)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"-12                       216.2m ±  2%   217.0m ±  9%         ~ (p=0.684 n=10)
Querier/Head/labelValuesWithMatchers/n_with_i="1"-12                          4.414µ ± 15%   4.309µ ±  4%         ~ (p=0.073 n=10)
Querier/Head/labelValuesWithMatchers/n_with_i=~"1.+"-12                       14.74m ±  3%   14.57m ±  2%         ~ (p=0.247 n=10)
Querier/Head/labelValuesWithMatchers/none_with_i=~"1"-12                  6948765.5n ± 57%   208.5n ±  1%  -100.00% (p=0.000 n=10)
geomean                                                                       15.56m         5.721m         -63.23%
```

I also was confused about this benchmark doing better:
```
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"-12                      148.97m ±  1%   42.59m ±  3%   -71.41% (p=0.000 n=10)
```

So I checked, and the benchmark was wrong: it's description should be:
```
Querier/Head/labelValuesWithMatchers/i_with_i="^.+$"
```

I pushed a commit fixing the label matcher on that benchmark (to actually match on `n`), but didn't run them again (as I assume it's unaffected by my change).